+++ kitty/boss.py
@@ -22,11 +22,12 @@ from .constants import (
 )
 from .fast_data_types import (
     ChildMonitor, background_opacity_of, change_background_opacity,
-    change_os_window_state, create_os_window, current_os_window,
-    destroy_global_data, focus_os_window, get_clipboard_string,
-    global_font_size, mark_os_window_for_close, os_window_font_size,
-    patch_global_colors, safe_pipe, set_background_image, set_clipboard_string,
-    set_in_sequence_mode, thread_write, toggle_fullscreen, toggle_maximized
+    change_os_window_state, cocoa_set_menubar_title, create_os_window,
+    current_os_window, destroy_global_data, focus_os_window,
+    get_clipboard_string, global_font_size, mark_os_window_for_close,
+    os_window_font_size, patch_global_colors, safe_pipe, set_background_image,
+    set_clipboard_string, set_in_sequence_mode, thread_write,
+    toggle_fullscreen, toggle_maximized
 )
 from .keys import get_shortcut, shortcut_matches
 from .layout import set_layout_options
@@ -627,6 +628,8 @@ class Boss:
             w = tm.active_window
             if w is not None:
                 w.focus_changed(focused)
+                if is_macos and focused:
+                    cocoa_set_menubar_title(w.title or '')
             tm.mark_tab_bar_dirty()
 
     def update_tab_bar_data(self, os_window_id):
@@ -648,6 +651,8 @@ class Boss:
             tm.destroy()
         for window_id in tuple(w.id for w in self.window_id_map.values() if getattr(w, 'os_window_id', None) == os_window_id):
             self.window_id_map.pop(window_id, None)
+        if not self.os_window_map and is_macos:
+            cocoa_set_menubar_title('')
         action = self.os_window_death_actions.pop(os_window_id, None)
         if action is not None:
             action()